# ddl: fast fail check for create index to accelerate CI