@tiancaiamao 2018-12-06
B.T.W, this change makes the session package test time reduce to 16.417s, from 53.792s